+++ src/main.zig
@@ -50,35 +50,32 @@ pub fn main() anyerror!void {
         .notes_dir = notes_dir,
     };
 
+    const builder = http.router.Builder(*Context);
+
     std.log.info("Serving on http://{s}:{}", .{ "127.0.0.1", DEFAULT_PORT });
     try http.listenAndServe(
         gpa.allocator(),
         try std.net.Address.parseIp("127.0.0.1", DEFAULT_PORT),
         &ctx,
-        index,
+        comptime http.router.Router(*Context, &.{
+            builder.get("/", null, index),
+            builder.post("/note", null, postNote),
+            builder.put("/folder/:folder/note", i64, putNoteInFolder),
+            builder.post("/folder", null, postFolder),
+            builder.get("/folder/:folder", i64, getFolder),
+            builder.post("/files/*", null, serveFs),
+            builder.get("/static/:filename", struct { filename: []const u8 }, staticFiles(.{
+                .@"style.css" = .{ .css = @embedFile("style.css") },
+                .@"tachyons.min.css" = .{ .css = @embedFile("tachyons.min.css") },
+                .@"htmx.min.js" = .{ .js = @embedFile("htmx.min.js") },
+                .@"_hyperscript.min.js" = .{ .js = @embedFile("_hyperscript.min.js") },
+            })),
+        }),
     );
 }
 
-fn index(ctx: *Context, res: *http.Response, req: http.Request) !void {
-    const path = req.path();
-    if (std.mem.startsWith(u8, path, "/files/")) {
-        return try http.FileServer.serve({}, res, req);
-    } else if (req.method() == .post and std.mem.eql(u8, path, "/note")) {
-        return postNote(ctx, res, req);
-    } else if (req.method() == .get and std.mem.eql(u8, path, "/static/htmx.min.js")) {
-        try res.headers.put("Content-Type", "text/javascript");
-        try res.writer().writeAll(@embedFile("./htmx.min.js"));
-        return;
-    } else if (req.method() == .get and std.mem.eql(u8, path, "/static/_hyperscript.min.js")) {
-        try res.headers.put("Content-Type", "text/javascript");
-        try res.writer().writeAll(@embedFile("./_hyperscript.min.js"));
-        return;
-    } else if (req.method() == .get and std.mem.eql(u8, path, "/static/tachyons.min.css")) {
-        try res.headers.put("Content-Type", "text/css");
-        try res.writer().writeAll(@embedFile("./tachyons.min.css"));
-        return;
-    }
-
+fn index(ctx: *Context, res: *http.Response, req: http.Request, _: ?*const anyopaque) !void {
+    _ = req;
     try res.headers.put("Content-Type", "text/html");
 
     var out = res.writer();
@@ -90,33 +87,75 @@ fn index(ctx: *Context, res: *http.Response, req: http.Request) !void {
         \\
         \\  <meta name="viewport" content="width=device-width, initial-scale=1">
         \\  <script src="/static/htmx.min.js"></script>
+        \\  <script src="/static/_hyperscript.min.js"></script>
         \\  <link rel="stylesheet" type="text/css" href="/static/tachyons.min.css">
+        \\  <link rel="stylesheet" type="text/css" href="/static/style.css">
         \\  <body>
+        \\    <div>
+        \\    <h2>Folders</h2>
+        \\    <form action="/folder" method="POST">
+        \\      <input type="text" name="name" />
+        \\      <button>New Folder</button>
+        \\    </form>
+        \\    <ul hx-include=".selected-note" id="folders" _="
+        \\        on dragover or dragenter halt the event
+        \\          then set the target's style.background to 'lightgray'
+        \\        on dragleave or drop set the target's style.background to ''
+        \\        on drop get event.dataTransfer.getData('my-notes-app/note-id')
+        \\          then put it into the next <output/>
+        \\    ">
+    );
+    {
+        var stmt = (try ctx.db.prepare_v2("SELECT id, name FROM folder", null)) orelse return error.NoStatement;
+        defer stmt.finalize() catch {};
+        while ((try stmt.step()) != .Done) {
+            const id = stmt.columnInt64(0);
+            const name = stmt.columnText(1);
+            // TODO: Escape note text
+            try out.print(
+                \\<li hx-put="/folder/{}/note" hx-trigger="drop" hx-swap="none" class="f3">
+                \\  <input type="hidden" name="folder_id" value="{}">
+                \\  <a class="no-underline" href="/folder/{}">{s}</a>
+                \\
+            , .{ id, id, id, name });
+        }
+    }
+    try out.writeAll(
+        \\    </ul>
+        \\    Dropped data: <output></output>
+        \\    </div>
+        \\    <div>
+        \\    <h2>Notes</h2>
         \\    <form action="/note" method="POST">
         \\      <input type="text" name="text" />
         \\      <button>Add note</button>
         \\    </form>
         \\    <div id="notes" class="pa1 flex flex-wrap justify-start" _="
-        \\        on dragstart call event.dataTransfer.setData('text/html', target.innerHTML)
+        \\        on dragstart
+        \\          add .selected-note to target
+        \\          call event.dataTransfer.setData('text/html', target.innerHTML)
+        \\          call event.dataTransfer.setData('text/plain', target.innerText)
+        \\          set id to <input[name='id']/> in event.target call event.dataTransfer.setData('my-notes-app/note-id', id.value)
         \\    ">
         \\
     );
-
-    var stmt = (try ctx.db.prepare_v2("SELECT id, text FROM note", null)) orelse return error.NoStatement;
-    defer stmt.finalize() catch {};
-    while ((try stmt.step()) != .Done) {
-        const id = stmt.columnInt64(0);
-        const text = stmt.columnText(1);
-        // TODO: Escape note text
-        try out.print(
-            \\<div class="w5 pa3 ma2 shadow-3" draggable="true">
-            \\  <input type="hidden" name="id" value="{}">
-            \\  {s}
-            \\</div>
-        , .{ id, text });
+    {
+        var stmt = (try ctx.db.prepare_v2("SELECT id, text FROM note WHERE folder_id = 0", null)) orelse return error.NoStatement;
+        defer stmt.finalize() catch {};
+        while ((try stmt.step()) != .Done) {
+            const id = stmt.columnInt64(0);
+            const text = stmt.columnText(1);
+            // TODO: Escape note text
+            try out.print(
+                \\<div class="w5 pa3 ma2 shadow-3" draggable="true">
+                \\  <input type="hidden" name="note-id" value="{}">
+                \\  {s}
+                \\</div>
+            , .{ id, text });
+        }
     }
-
     try out.writeAll(
+        \\    </div>
         \\    </div>
         \\  </body>
         \\</html>
@@ -124,7 +163,98 @@ fn index(ctx: *Context, res: *http.Response, req: http.Request) !void {
     );
 }
 
-fn postNote(ctx: *Context, res: *http.Response, req: http.Request) !void {
+fn getFolder(ctx: *Context, res: *http.Response, req: http.Request, captures: ?*const anyopaque) !void {
+    _ = req;
+    const folder_id = @ptrCast(?*const i64, @alignCast(@alignOf(?*const i64), captures)) orelse return error.hell;
+    try res.headers.put("Content-Type", "text/html");
+
+    var out = res.writer();
+    try out.writeAll(
+        \\<!DOCTYPE html>
+        \\<html lang="en">
+        \\  <meta charset="utf-8">
+        \\  <title>My Notes App</title>
+        \\
+        \\  <meta name="viewport" content="width=device-width, initial-scale=1">
+        \\  <script src="/static/htmx.min.js"></script>
+        \\  <script src="/static/_hyperscript.min.js"></script>
+        \\  <link rel="stylesheet" type="text/css" href="/static/tachyons.min.css">
+        \\  <link rel="stylesheet" type="text/css" href="/static/style.css">
+        \\  <body>
+        \\    <div>
+        \\    <h2>Folders</h2>
+        \\    <form action="/folder" method="POST">
+        \\      <input type="text" name="name" />
+        \\      <button>New Folder</button>
+        \\    </form>
+        \\    <ul hx-include=".selected-note" id="folders" _="
+        \\        on dragover or dragenter halt the event
+        \\          then set the target's style.background to 'lightgray'
+        \\        on dragleave or drop set the target's style.background to ''
+        \\        on drop get event.dataTransfer.getData('my-notes-app/note-id')
+        \\          then put it into the next <output/>
+        \\    ">
+    );
+    {
+        var stmt = (try ctx.db.prepare_v2("SELECT id, name FROM folder", null)) orelse return error.NoStatement;
+        defer stmt.finalize() catch {};
+        while ((try stmt.step()) != .Done) {
+            const id = stmt.columnInt64(0);
+            const name = stmt.columnText(1);
+            // TODO: Escape note text
+            try out.print(
+                \\<li hx-put="/folder/{}/note" hx-trigger="drop" hx-swap="none" class="f3">
+                \\  <input type="hidden" name="folder_id" value="{}">
+                \\  <a class="no-underline" href="/folder/{}">{s}</a>
+                \\
+            , .{ id, id, id, name });
+        }
+    }
+    try out.writeAll(
+        \\    </ul>
+        \\    Dropped data: <output></output>
+        \\    </div>
+        \\    <div>
+        \\    <h2>Notes</h2>
+        \\    <form action="/note" method="POST">
+        \\      <input type="text" name="text" />
+        \\      <button>Add note</button>
+        \\    </form>
+        \\    <div id="notes" class="pa1 flex flex-wrap justify-start" _="
+        \\        on dragstart
+        \\          add .selected-note to target
+        \\          call event.dataTransfer.setData('text/html', target.innerHTML)
+        \\          call event.dataTransfer.setData('text/plain', target.innerText)
+        \\          set id to <input[name='id']/> in event.target call event.dataTransfer.setData('my-notes-app/note-id', id.value)
+        \\    ">
+        \\
+    );
+    {
+        var stmt = (try ctx.db.prepare_v2("SELECT id, text FROM note WHERE folder_id = ?", null)) orelse return error.NoStatement;
+        defer stmt.finalize() catch {};
+        try stmt.bindInt64(1, folder_id.*);
+        while ((try stmt.step()) != .Done) {
+            const id = stmt.columnInt64(0);
+            const text = stmt.columnText(1);
+            // TODO: Escape note text
+            try out.print(
+                \\<div class="w5 pa3 ma2 shadow-3" draggable="true">
+                \\  <input type="hidden" name="note-id" value="{}">
+                \\  {s}
+                \\</div>
+            , .{ id, text });
+        }
+    }
+    try out.writeAll(
+        \\    </div>
+        \\    </div>
+        \\  </body>
+        \\</html>
+        \\
+    );
+}
+
+fn postNote(ctx: *Context, res: *http.Response, req: http.Request, _: ?*const anyopaque) !void {
     const text = (try req.formValue(ctx.allocator, "text")) orelse {
         return error.InvalidInput;
     };
@@ -151,6 +281,72 @@ fn postNote(ctx: *Context, res: *http.Response, req: http.Request) !void {
         \\</html>
         \\
     );
+}
+
+fn postFolder(ctx: *Context, res: *http.Response, req: http.Request, _: ?*const anyopaque) !void {
+    const name = (try req.formValue(ctx.allocator, "name")) orelse {
+        return error.InvalidInput;
+    };
+    defer ctx.allocator.free(name);
+
+    var stmt = (try ctx.db.prepare_v2("INSERT INTO folder(name) VALUES (?)", null)) orelse return error.NoStatement;
+    defer stmt.finalize() catch {};
+    try stmt.bindText(1, name, .transient);
+    if ((try stmt.step()) != .Done) {
+        return error.UnexpectedReturnCode;
+    }
+
+    res.status_code = .see_other;
+    try res.headers.put("Location", "/");
+    try res.headers.put("Content-Type", "text/html");
+
+    var out = res.writer();
+    try out.writeAll(
+        \\<!DOCTYPE html>
+        \\<html>
+        \\<body>
+        \\Folder posted
+        \\</body>
+        \\</html>
+        \\
+    );
+}
+
+fn putNoteInFolder(ctx: *Context, res: *http.Response, req: http.Request, captures: ?*const anyopaque) !void {
+    const folder_id = (@ptrCast(?*const i64, @alignCast(@alignOf(?*const i64), captures)) orelse return error.hell).*;
+    const note_id_str = (try req.formValue(ctx.allocator, "note-id")) orelse {
+        std.debug.print("form value = {}\n", .{std.zig.fmtEscapes(req.body())});
+        return error.InvalidInput;
+    };
+    defer ctx.allocator.free(note_id_str);
+
+    const note_id = try std.fmt.parseInt(i64, note_id_str, 10);
+
+    var txn = try Transaction.begin(@src(), ctx.db);
+    defer txn.deinit();
+
+    var stmt = (try ctx.db.prepare_v2("UPDATE note SET folder_id = ? WHERE id = ?", null)) orelse return error.NoStatement;
+    defer stmt.finalize() catch {};
+    try stmt.bindInt64(1, folder_id);
+    try stmt.bindInt64(2, note_id);
+    while ((try stmt.step()) != .Done) {}
+
+    res.status_code = .see_other;
+    try res.headers.put("Location", "/");
+    try res.headers.put("Content-Type", "text/html");
+
+    var out = res.writer();
+    try out.writeAll(
+        \\<!DOCTYPE html>
+        \\<html>
+        \\<body>
+        \\Note moved to folder
+        \\</body>
+        \\</html>
+        \\
+    );
+
+    try txn.commit();
 }
 
 /// - https://david.rothlis.net/declarative-schema-migration-for-sqlite/
@@ -180,15 +376,12 @@ fn setupSchema(allocator: std.mem.Allocator, db: *sqlite3.SQLite3) !void {
             // The table already exists
             if (current_table_info.eql(pristine_table_entry.value_ptr.*)) {
                 // The tables are identical, nothing needs to change
-                std.debug.print("not adding table: {}\n", .{std.zig.fmtEscapes(pristine_table_entry.key_ptr.*)});
                 continue;
             }
             // The tables are not identical, we need to rebuild it
-            std.debug.print("rebuilding table: {}\n", .{std.zig.fmtEscapes(pristine_table_entry.key_ptr.*)});
             try dbRebuildTable(allocator, db, pristine, pristine_table_entry.value_ptr.*);
             continue;
         }
-        std.debug.print("adding table: {}\n", .{std.zig.fmtEscapes(pristine_table_entry.key_ptr.*)});
         // Create table in current database that is in pristine database
         try db.exec(pristine_table_entry.value_ptr.sql, null, null, null);
     }
@@ -198,14 +391,14 @@ fn setupSchema(allocator: std.mem.Allocator, db: *sqlite3.SQLite3) !void {
     while (current_table_iter.next()) |current_table_entry| {
         if (pristine_tables.contains(current_table_entry.key_ptr.*)) {
             // The table exists in the pristine, no need to drop it
-            std.debug.print("not dropping table: {}\n", .{std.zig.fmtEscapes(current_table_entry.key_ptr.*)});
             continue;
         }
-        std.debug.print("dropping table: {}\n", .{std.zig.fmtEscapes(current_table_entry.key_ptr.*)});
         // Drop table in current database that is not in pristine database
         const drop_sql = try std.fmt.allocPrintZ(arena.allocator(), "DROP TABLE '{}'", .{std.zig.fmtEscapes(current_table_entry.key_ptr.*)});
         try db.exec(drop_sql, null, null, null);
     }
+
+    try db.exec("INSERT OR IGNORE INTO folder (id, name) VALUES (0, 'Inbox')", null, null, null);
 
     try txn.commit();
 }
@@ -239,7 +432,6 @@ const TableInfo = struct {
 
 /// - sqlite 3.37 introduced `PRAGMA table_list`: https://sqlite.org/pragma.html#pragma_table_list
 fn dbTables(allocator: std.mem.Allocator, db: *sqlite3.SQLite3) !std.StringHashMap(TableInfo) {
-    std.debug.print("{s}:{} hello\n", .{ @src().file, @src().line });
     var stmt = (try db.prepare_v2(
         \\ SELECT tl.schema, tl.name, tl.type, tl.ncol, tl.wr, tl.strict, schema.sql
         \\ FROM pragma_table_list AS tl
@@ -247,7 +439,6 @@ fn dbTables(allocator: std.mem.Allocator, db: *sqlite3.SQLite3) !std.StringHashM
     , null)).?;
     defer stmt.finalize() catch unreachable;
 
-    std.debug.print("{s}:{} hello\n", .{ @src().file, @src().line });
     var hashmap = std.StringHashMap(TableInfo).init(allocator);
     errdefer hashmap.deinit();
     while ((try stmt.step()) != .Done) {
@@ -260,17 +451,6 @@ fn dbTables(allocator: std.mem.Allocator, db: *sqlite3.SQLite3) !std.StringHashM
         const sql = try allocator.dupeZ(u8, stmt.columnText(6));
 
         const fullname = try std.fmt.allocPrint(allocator, "\"{}\".\"{}\"", .{ std.zig.fmtEscapes(schema), std.zig.fmtEscapes(name) });
-        std.debug.print("{s}:{} {}.{} = {s} {} {} {s} \"{}\"\n", .{
-            @src().file,
-            @src().line,
-            std.zig.fmtEscapes(schema),
-            std.zig.fmtEscapes(name),
-            table_type,
-            ncol,
-            without_rowid,
-            if (strict) @as([]const u8, "STRICT") else "",
-            std.zig.fmtEscapes(sql),
-        });
         try hashmap.putNoClobber(fullname, .{
             .schema = schema,
             .name = name,
@@ -282,54 +462,6 @@ fn dbTables(allocator: std.mem.Allocator, db: *sqlite3.SQLite3) !std.StringHashM
         });
     }
 
-    std.debug.print("{s}:{} hello\n", .{ @src().file, @src().line });
-    return hashmap;
-}
-
-/// - sqlite 3.37 introduced `PRAGMA table_list`: https://sqlite.org/pragma.html#pragma_table_list
-fn dbTableInfo(allocator: std.mem.Allocator, db: *sqlite3.SQLite3) !std.StringArrayHashMap(TableInfo) {
-    std.debug.print("{s}:{} hello\n", .{ @src().file, @src().line });
-    var stmt = (try db.prepare_v2(
-        \\ SELECT name FROM pragma_table_info(?)
-    , null)).?;
-    defer stmt.finalize() catch unreachable;
-
-    std.debug.print("{s}:{} hello\n", .{ @src().file, @src().line });
-    var hashmap = std.StringHashMap(TableInfo).init(allocator);
-    errdefer hashmap.deinit();
-    while ((try stmt.step()) != .Done) {
-        const schema = try allocator.dupeZ(u8, stmt.columnText(0));
-        const name = try allocator.dupeZ(u8, stmt.columnText(1));
-        const table_type = std.meta.stringToEnum(TableInfo.Type, stmt.columnText(2)).?;
-        const ncol = @intCast(u32, stmt.columnInt(3));
-        const without_rowid = stmt.columnInt(4) != 0;
-        const strict = stmt.columnInt(5) != 0;
-        const sql = try allocator.dupeZ(u8, stmt.columnText(6));
-
-        const fullname = try std.fmt.allocPrint(allocator, "\"{}\".\"{}\"", .{ std.zig.fmtEscapes(schema), std.zig.fmtEscapes(name) });
-        std.debug.print("{s}:{} {}.{} = {s} {} {} {s} \"{}\"\n", .{
-            @src().file,
-            @src().line,
-            std.zig.fmtEscapes(schema),
-            std.zig.fmtEscapes(name),
-            table_type,
-            ncol,
-            without_rowid,
-            if (strict) @as([]const u8, "STRICT") else "",
-            std.zig.fmtEscapes(sql),
-        });
-        try hashmap.putNoClobber(fullname, .{
-            .schema = schema,
-            .name = name,
-            .table_type = table_type,
-            .ncol = ncol,
-            .without_rowid = without_rowid,
-            .strict = strict,
-            .sql = sql,
-        });
-    }
-
-    std.debug.print("{s}:{} hello\n", .{ @src().file, @src().line });
     return hashmap;
 }
 
@@ -422,6 +554,8 @@ fn dbRebuildTable(allocator: std.mem.Allocator, db: *sqlite3.SQLite3, pristine: 
     try db.exec(alter_sql, null, null, null);
 
     try db.exec("PRAGMA foreign_keys_check;", null, null, null);
+
+    try txn.commit();
 }
 
 fn executeScript(db: *sqlite3.SQLite3, sql: []const u8) !void {
@@ -429,12 +563,6 @@ fn executeScript(db: *sqlite3.SQLite3, sql: []const u8) !void {
     var tail_sql = sql;
     while (try db.prepare_v2(next_sql, &tail_sql)) |stmt| : (next_sql = tail_sql) {
         defer stmt.finalize() catch |e| std.debug.panic("could not finalize: {}", .{e});
-        std.debug.print(
-            \\```sql
-            \\{s}
-            \\```
-            \\
-        , .{next_sql[0..@ptrToInt(tail_sql.ptr) -| @ptrToInt(next_sql.ptr)]});
         while ((try stmt.step()) != .Done) {}
     }
 }
@@ -471,6 +599,59 @@ const Transaction = struct {
         }
     }
 };
+
+pub const Blob = union(enum) {
+    html: []const u8,
+    css: []const u8,
+    js: []const u8,
+
+    pub fn contentType(this: @This()) []const u8 {
+        return switch (this) {
+            .html => "text/html",
+            .css => "text/css",
+            .js => "text/javascript",
+        };
+    }
+
+    pub fn data(this: @This()) []const u8 {
+        return switch (this) {
+            .html => |d| d,
+            .css => |d| d,
+            .js => |d| d,
+        };
+    }
+};
+
+/// Expects an struct where the field names are files, and the values are Blobs
+fn staticFiles(comptime static_files: anytype) fn (*Context, *http.Response, http.Request, ?*const anyopaque) http.Response.Error!void {
+    const Handler = struct {
+        fn handle(_: *Context, res: *http.Response, req: http.Request, captures: ?*const anyopaque) http.Response.Error!void {
+            _ = req;
+
+            if (captures == null) return res.notFound();
+
+            const req_filename = @ptrCast(*const struct { filename: []const u8 }, @alignCast(@alignOf(Blob), captures.?)).filename;
+            inline for (std.meta.fields(@TypeOf(static_files))) |field| {
+                if (std.mem.eql(u8, req_filename, field.name)) {
+                    const static_file: Blob = @field(static_files, field.name);
+                    try res.headers.put("Content-Type", static_file.contentType());
+                    try res.writer().writeAll(static_file.data());
+                    return;
+                }
+            }
+
+            return res.notFound();
+        }
+    };
+    return Handler.handle;
+}
+
+/// Serves a file
+fn serveFs(ctx: *Context, resp: *http.Response, req: http.Request, captures: ?*const anyopaque) !void {
+    std.debug.assert(captures == null);
+    _ = ctx;
+    try http.FileServer.serve({}, resp, req);
+}
 
 fn srcLineStr(comptime src: std.builtin.SourceLocation) *const [srcLineStrLen(src):0]u8 {
     return std.fmt.comptimePrint("{s}:{}", .{ src.file, src.line });